poprawiony algorytm losowania

+++ Zgadywanie_liczb/app/src/main/java/com/example/zgadywanie_liczb/Counter.java
@@ -28,7 +28,7 @@ public void  setHighScore(int newScore){
 
     public double calculateHighScore(int score, int range){
         double result;
-        result = (score / range) * 10000;
+        result = ((double)score / (double)range) * 10000d;
         return result;
     }
